backend/requirements.txt
fastapi
uvicorn
pydantic
xhtml2pdf
jinja2
python-multipart
langgraph
langchain
langchain-openai
markdown
python-dotenv
